html_problem_selection(): respect hide_variant
This effectively hides variant number in the clar request
submission interface if hide_variant is set. (See the use of
the changed function in unpriv_main_clar_submit.csp.)

I think, html_problem_selection_2() should be changed in the
same way, but currently don't have a setup to test that.