Import ML as max_rss_size from Polygon
This is a new, more correct replacement for the max_vm_size option.
Using max_rss_size by default avoids difficulties with runtimes and
libraries that allocate large heaps or (ab)use the page table structures
as an efficient trie. This is also consistent with Yandex.Contest and
most cgroups-based test systems.